refactor(lowy): extract a shared Commit renderer
The hash→GitHub-link logic lived in both IdentityRail's local `Commit`
and an inline block in App.tsx. The two had drifted: App.tsx guarded
only `!== "dev"`, so a dirty build (`<sha>-dirty`) rendered a broken
/commit/ link. Extract one `ui/Commit.tsx` (the stricter clean-ref
guard) used by both the rail (server/pty commit) and the About dialog
(client build commit) — distinct data sources, one renderer.
